Add tracks to ABTest package
This adds two tracks events to ABTest package. The first when a group is
assigned (`abtest_assign_variant`) and the second when a variant is
served (`abtest_serve_variant`).

The events both send the following props:

abtest_name - name of test
abtest_group - assigned group
context - additional context for test